requirements.txt
python-dotenv==1.0.0
anthropic==0.25.0
supabase==2.5.0
pyyaml==6.0.1
requests==2.31.0
httpx>=0.24.0
pytest>=7.4.0
pytest-asyncio>=1.3.0
playwright>=1.40.0